requirements.txt
fastapi>=0.115.12
uvicorn>=0.29.0
aiohttp>=3.12.15
pydantic>=2.11.7
python-dotenv>=1.1.1
playwright>=1.54.0
nest-asyncio>=1.6.0
psutil>=7.0.0